+++ packages/indexer/src/main/kotlin/org/vechain/indexer/migration/v1_0_0/TransferEventsInitializerChangeUnit.kt
@@ -21,7 +21,7 @@ class TransferEventsInitializerChangeUnit {
 
     @BeforeExecution
     fun beforeExecution(mongoTemplate: MongoTemplate) {
-        mongoTemplate.createCollection(TRANSFER_EVENTS)
+        if (!mongoTemplate.collectionExists(TRANSFER_EVENTS)) mongoTemplate.createCollection(TRANSFER_EVENTS)
     }
 
     @Execution
@@ -75,6 +75,6 @@ class TransferEventsInitializerChangeUnit {
 
     @RollbackBeforeExecution
     fun rollbackBeforeExecution(mongoTemplate: MongoTemplate) {
-        mongoTemplate.dropCollection(TRANSFER_EVENTS)
+        if (mongoTemplate.collectionExists(TRANSFER_EVENTS)) mongoTemplate.dropCollection(TRANSFER_EVENTS)
     }
 }
